Moved cast to own module.
