+++ rust/datafusion/src/physical_plan/expressions/mod.rs
@@ -32,7 +32,6 @@ use arrow::array::{
     StringOffsetSizeTrait,
 };
 use arrow::compute;
-use arrow::compute::kernels;
 use arrow::compute::kernels::arithmetic::{add, divide, multiply, negate, subtract};
 use arrow::compute::kernels::boolean::{and, nullif, or};
 use arrow::compute::kernels::comparison::{eq, gt, gt_eq, lt, lt_eq, neq};
@@ -58,9 +57,10 @@ use arrow::{
     },
     datatypes::Field,
 };
-use compute::can_cast_types;
 
+mod cast;
 mod coercion;
+pub use cast::{cast, CastExpr};
 
 /// returns the name of the state
 pub fn format_state_name(name: &str, state_name: &str) -> String {
@@ -2226,91 +2226,6 @@ impl PhysicalExpr for CaseExpr {
     }
 }
 
-/// CAST expression casts an expression to a specific data type
-#[derive(Debug)]
-pub struct CastExpr {
-    /// The expression to cast
-    expr: Arc<dyn PhysicalExpr>,
-    /// The data type to cast to
-    cast_type: DataType,
-}
-
-impl CastExpr {
-    /// Create a new CastExpr
-    pub fn new(expr: Arc<dyn PhysicalExpr>, cast_type: DataType) -> Self {
-        Self { expr, cast_type }
-    }
-
-    /// The expression to cast
-    pub fn expr(&self) -> &Arc<dyn PhysicalExpr> {
-        &self.expr
-    }
-
-    /// The data type to cast to
-    pub fn cast_type(&self) -> &DataType {
-        &self.cast_type
-    }
-}
-
-impl fmt::Display for CastExpr {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "CAST({} AS {:?})", self.expr, self.cast_type)
-    }
-}
-
-impl PhysicalExpr for CastExpr {
-    /// Return a reference to Any that can be used for downcasting
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-
-    fn data_type(&self, _input_schema: &Schema) -> Result<DataType> {
-        Ok(self.cast_type.clone())
-    }
-
-    fn nullable(&self, input_schema: &Schema) -> Result<bool> {
-        self.expr.nullable(input_schema)
-    }
-
-    fn evaluate(&self, batch: &RecordBatch) -> Result<ColumnarValue> {
-        let value = self.expr.evaluate(batch)?;
-        match value {
-            ColumnarValue::Array(array) => Ok(ColumnarValue::Array(kernels::cast::cast(
-                &array,
-                &self.cast_type,
-            )?)),
-            ColumnarValue::Scalar(scalar) => {
-                let scalar_array = scalar.to_array();
-                let cast_array = kernels::cast::cast(&scalar_array, &self.cast_type)?;
-                let cast_scalar = ScalarValue::try_from_array(&cast_array, 0)?;
-                Ok(ColumnarValue::Scalar(cast_scalar))
-            }
-        }
-    }
-}
-
-/// Return a PhysicalExpression representing `expr` casted to
-/// `cast_type`, if any casting is needed.
-///
-/// Note that such casts may lose type information
-pub fn cast(
-    expr: Arc<dyn PhysicalExpr>,
-    input_schema: &Schema,
-    cast_type: DataType,
-) -> Result<Arc<dyn PhysicalExpr>> {
-    let expr_type = expr.data_type(input_schema)?;
-    if expr_type == cast_type {
-        Ok(expr.clone())
-    } else if can_cast_types(&expr_type, &cast_type) {
-        Ok(Arc::new(CastExpr::new(expr, cast_type)))
-    } else {
-        Err(DataFusionError::Internal(format!(
-            "Unsupported CAST from {:?} to {:?}",
-            expr_type, cast_type
-        )))
-    }
-}
-
 /// Represents a non-null literal value
 #[derive(Debug)]
 pub struct Literal {
@@ -2646,7 +2561,7 @@ mod tests {
     use arrow::{
         array::{
             LargeStringArray, PrimitiveArray, PrimitiveBuilder, StringArray,
-            StringDictionaryBuilder, Time64NanosecondArray,
+            StringDictionaryBuilder,
         },
         util::display::array_value_to_string,
     };
@@ -2993,103 +2908,6 @@ mod tests {
                 "Coercion should have returned an DataFusionError::Internal".to_string(),
             ))
         }
-    }
-
-    // runs an end-to-end test of physical type cast
-    // 1. construct a record batch with a column "a" of type A
-    // 2. construct a physical expression of CAST(a AS B)
-    // 3. evaluate the expression
-    // 4. verify that the resulting expression is of type B
-    // 5. verify that the resulting values are downcastable and correct
-    macro_rules! generic_test_cast {
-        ($A_ARRAY:ident, $A_TYPE:expr, $A_VEC:expr, $TYPEARRAY:ident, $TYPE:expr, $VEC:expr) => {{
-            let schema = Schema::new(vec![Field::new("a", $A_TYPE, false)]);
-            let a = $A_ARRAY::from($A_VEC);
-            let batch =
-                RecordBatch::try_new(Arc::new(schema.clone()), vec![Arc::new(a)])?;
-
-            // verify that we can construct the expression
-            let expression = cast(col("a"), &schema, $TYPE)?;
-
-            // verify that its display is correct
-            assert_eq!(format!("CAST(a AS {:?})", $TYPE), format!("{}", expression));
-
-            // verify that the expression's type is correct
-            assert_eq!(expression.data_type(&schema)?, $TYPE);
-
-            // compute
-            let result = expression.evaluate(&batch)?.into_array(batch.num_rows());
-
-            // verify that the array's data_type is correct
-            assert_eq!(*result.data_type(), $TYPE);
-
-            // verify that the len is correct
-            assert_eq!(result.len(), $A_VEC.len());
-
-            // verify that the data itself is downcastable
-            let result = result
-                .as_any()
-                .downcast_ref::<$TYPEARRAY>()
-                .expect("failed to downcast");
-
-            // verify that the result itself is correct
-            for (i, x) in $VEC.iter().enumerate() {
-                assert_eq!(result.value(i), *x);
-            }
-        }};
-    }
-
-    #[test]
-    fn test_cast_i32_u32() -> Result<()> {
-        generic_test_cast!(
-            Int32Array,
-            DataType::Int32,
-            vec![1, 2, 3, 4, 5],
-            UInt32Array,
-            DataType::UInt32,
-            vec![1_u32, 2_u32, 3_u32, 4_u32, 5_u32]
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn test_cast_i32_utf8() -> Result<()> {
-        generic_test_cast!(
-            Int32Array,
-            DataType::Int32,
-            vec![1, 2, 3, 4, 5],
-            StringArray,
-            DataType::Utf8,
-            vec!["1", "2", "3", "4", "5"]
-        );
-        Ok(())
-    }
-    #[allow(clippy::redundant_clone)]
-    #[test]
-    fn test_cast_i64_t64() -> Result<()> {
-        let original = vec![1, 2, 3, 4, 5];
-        let expected: Vec<i64> = original
-            .iter()
-            .map(|i| Time64NanosecondArray::from(vec![*i]).value(0))
-            .collect();
-        generic_test_cast!(
-            Int64Array,
-            DataType::Int64,
-            original.clone(),
-            TimestampNanosecondArray,
-            DataType::Timestamp(TimeUnit::Nanosecond, None),
-            expected
-        );
-        Ok(())
-    }
-
-    #[test]
-    fn invalid_cast() -> Result<()> {
-        // Ensure a useful error happens at plan time if invalid casts are used
-        let schema = Schema::new(vec![Field::new("a", DataType::Int32, false)]);
-        let result = cast(col("a"), &schema, DataType::LargeBinary);
-        result.expect_err("expected Invalid CAST");
-        Ok(())
     }
 
     /// macro to perform an aggregation and verify the result.
